Split deprecations up to make ongoing management easier
