+++ packages/block-library/src/cover/deprecated/shared.js
@@ -1,0 +1,85 @@
+const POSITION_CLASSNAMES = {
+	'top left': 'is-position-top-left',
+	'top center': 'is-position-top-center',
+	'top right': 'is-position-top-right',
+	'center left': 'is-position-center-left',
+	'center center': 'is-position-center-center',
+	center: 'is-position-center-center',
+	'center right': 'is-position-center-right',
+	'bottom left': 'is-position-bottom-left',
+	'bottom center': 'is-position-bottom-center',
+	'bottom right': 'is-position-bottom-right',
+};
+
+export const IMAGE_BACKGROUND_TYPE = 'image';
+export const VIDEO_BACKGROUND_TYPE = 'video';
+
+export const blockAttributes = {
+	url: {
+		type: 'string',
+	},
+	id: {
+		type: 'number',
+	},
+	hasParallax: {
+		type: 'boolean',
+		default: false,
+	},
+	dimRatio: {
+		type: 'number',
+		default: 50,
+	},
+	overlayColor: {
+		type: 'string',
+	},
+	customOverlayColor: {
+		type: 'string',
+	},
+	backgroundType: {
+		type: 'string',
+		default: 'image',
+	},
+	focalPoint: {
+		type: 'object',
+	},
+};
+
+export function dimRatioToClass( ratio ) {
+	return ratio === 0 || ratio === 50 || ! ratio
+		? null
+		: 'has-background-dim-' + 10 * Math.round( ratio / 10 );
+}
+
+export function backgroundImageStyles( url ) {
+	return url ? { backgroundImage: `url(${ url })` } : {};
+}
+
+/**
+ * Checks of the contentPosition is the center (default) position.
+ *
+ * @param {string} contentPosition The current content position.
+ * @return {boolean} Whether the contentPosition is center.
+ */
+export function isContentPositionCenter( contentPosition ) {
+	return (
+		! contentPosition ||
+		contentPosition === 'center center' ||
+		contentPosition === 'center'
+	);
+}
+
+/**
+ * Retrieves the className for the current contentPosition.
+ * The default position (center) will not have a className.
+ *
+ * @param {string} contentPosition The current content position.
+ * @return {string} The className assigned to the contentPosition.
+ */
+export function getPositionClassName( contentPosition ) {
+	/*
+	 * Only render a className if the contentPosition is not center (the default).
+	 */
+	if ( isContentPositionCenter( contentPosition ) ) return '';
+
+	return POSITION_CLASSNAMES[ contentPosition ];
+}
